+++ test/test-runner.ts
@@ -330,3 +330,29 @@ test('tryUntil', async (t) => {
         'shall abort if tryUntil fails'
     )
 })
+
+test('error-comment', async (t) => {
+    const res = await runTest('./error-comment.js')
+    const testsuites = res.xml.testsuites
+
+    t.equal(res.exitCode, 1, 'test shall not pass')
+    t.equal(testsuites.$.tests, '6')
+    t.equal(testsuites.$.errors, '1')
+    t.equal(testsuites.$.failures, '1')
+    t.equal(testsuites.testsuite[0].testcase[1].$.status, 'success')
+    t.equal(testsuites.testsuite[0].testcase[1].$.assertions, '1')
+    t.equal(testsuites.testsuite[0].testcase[3].$.status, 'failed')
+    t.equal(testsuites.testsuite[0].testcase[3].$.assertions, '1')
+    t.equal(
+        testsuites.testsuite[0].testcase[3].failure[0]['_'],
+        'Fails\ncomment1\ncomment2'
+    )
+    t.equal(testsuites.testsuite[0].testcase[5].$.status, 'error')
+    t.equal(testsuites.testsuite[0].testcase[3].$.assertions, '1')
+    t.true(
+        testsuites.testsuite[0].testcase[3].failure[0]['_'].match(
+            '\ncomment1\ncomment2'
+        )
+    )
+    t.equal(testsuites.$.skipped, '0')
+})